Split the lib.rs file into smaller, more manageable submodules.
